Commit of working code - c++ multidispatcher
First commit for code that works, and does the job okay. This code
uses a combination of openMP and the system() call to launch
a load of parallel jobs. The jobs should be specified in a file,
the name of which should be given on the command line. The number of
processes to launch is the second argument.

This should work with any program that works/has been compiled
for the system you're running on. Once I figure out the flags
and defines I need, I'll re-write to dispatch on XeonPhi, to each
card individually. Hopefully this will be as easy as the intel docs
claim...